Work on cleaning up resource related TODOs: - Make ChemicalStacks use ints for amounts to avoid needing to clamp the value in a variety of places - Create sync classes for large resource stacks, and use them to sync fluid and chemical tanks that might have a size larger than max int - Expose resources to computers - Deduplicate some code between chemicals and fluids for bar/gauge GUI elements

+++ src/main/java/mekanism/client/recipe_viewer/emi/recipe/MekanismEmiRecipe.java
@@ -255,7 +255,7 @@ public abstract class MekanismEmiRecipe<RECIPE> extends AbstractContainerEventHa
         }
         List<ChemicalEmiStack> list = new ArrayList<>(representations.size());
         for (ChemicalStack stack : representations) {
-            list.add(new ChemicalEmiStack(stack.typeHolder(), stack.amount() * scalar));
+            list.add(new ChemicalEmiStack(stack.typeHolder(), (long) stack.amount() * scalar));
         }
         return EmiIngredient.of(list);
     }
